Migrate Java 10 tests out of supersource
